kyme/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
itsdangerous==2.1.2
kubernetes==29.0.0
jsonschema==4.21.0
